requirements.txt
Keras>=2.3.1
matlabengineforpython>=R2015b
numpy>=1.17.4
tensorflow>=2.0.0
